test: Make many percolator integration tests real integration tests
